Consolidate scene actions.
