+++ src/js/components/visualization/LayerList.jsx
@@ -10,7 +10,7 @@ var React = require('react'),
     inspectorActions = require('../../actions/inspectorActions'),
     selectMark = inspectorActions.selectMark,
     expandLayers = inspectorActions.expandLayers,
-    clearScene = require('../../actions/clearScene'),
+    clearScene = require('../../actions/sceneActions').clearScene,
     Group = require('./GroupSubMenu'),
     assets = require('../../util/assets'),
     Icon = require('../Icon');
