+++ src/js/reducers/signals.test.js
@@ -6,7 +6,7 @@ var Immutable = require('immutable');
 var actions = require('../constants/actions');
 var signalsReducer = require('./signals');
 var markActions = require('../actions/markActions');
-var createScene = require('../actions/createScene');
+var createScene = require('../actions/sceneActions').createScene;
 var signalActions = require('../actions/signalActions');
 var setSignalStreams = signalActions.setSignalStreams;
 var initSignal = signalActions.initSignal;
